get cli port from config
