+++ arduino-ide-extension/src/node/config-service-impl.ts
@@ -216,7 +216,7 @@ export class ConfigServiceImpl
   protected async mapCliConfigToAppConfig(
     cliConfig: DefaultCliConfig
   ): Promise<Config> {
-    const { directories, locale = 'en' } = cliConfig;
+    const { directories, locale = 'en', daemon } = cliConfig;
     const { data, user, downloads } = directories;
     const additionalUrls: Array<string> = [];
     if (cliConfig.board_manager && cliConfig.board_manager.additional_urls) {
@@ -232,6 +232,7 @@ export class ConfigServiceImpl
       additionalUrls,
       network,
       locale,
+      daemon,
     };
   }
